OSRA: handle identical-but-negated conditions
If a conditional branch propagates a constraint on a value that is
exactly the opposite with respect to its current constraint, we simply
ignore it, since simply flipping the condition (and the destination
basic blocks) would do the same.

In future, we should propagate a contradiction on the appropriate
branch.