OSRA: improve handling of equality comparisons
Equality comparisons used to be ignored more often than required due to
the fact that they give no hint on the signedness of the tracked value.

This commits removes some assertions and improves the handling of OSRs
without a known signedness. In particular, now, inequalities can be
solved even in absence of signedness information, as long as the result
that you would get with a signed OSR and an unsigned OSR matches (i.e.,
the signedness doesn't matter).