Add argument to `InlineAsm` function
`InlineAsm` instructions resemble function calls. You also have to
provide the prototype of the called function. Before this commit, we
were employing `void(void)`, which was wrong since we were passing in a
CSV.

In this commit we add a pointer argument to this prototype, which
prevents an assertion in LLVM from being triggered.