Being a callee is more important than being read
We used to black list memory portions target of memory read, so that
they couldn't become function entry points. However, this led to false
positives, in particular with functions that are explicitly called.

This commit ensures that we black list those addresses only if they are
not targets of function calls.