Improved SCS identification phase
Now in the `SCS` identification phase, when creating a SCS, if we
contain a node which is the source of another (sub)SCS, we also include
that SCS in the current SCS (in order to increase the number of nested
subregions).

The exception to this situation is when the source of the
other SCS is the head of the region itself under analysis.