Avoid emission of duplicate functions
In the Decompiler::HandleTranslationUnit method we abuse clang::tooling
to call clang on an empty file, each time a new Function is decompiled.
To do this we use clang::tooling's CommonOptionParser, which appends the
names of the parsed files to a static cl::list. The fact that this list
is static, means that it is never cleaned between calls to
DecompilationPass, which means that at each successive call
clang::tooling thinks we have more and more input files, meaning that it
creates multiple translation units. For each translation unit we run a
DecompilationAction, hence we emit multiple identical copies of the same
function. To disable this we have to create CommonOptionParser only
once, making it static as well.