New SCS merge step for abnormal retreating.
New SCS merge step for situations in which we have an abnormal edge
exiting from an SCS.

This criterion, during the preprocessing of the metaregions, if an SCS
with an abnormal retreating edge is found (an abnormal retreating is a
retreating edge which target is outside the scope of the current SCS
under analysis), merges the metaregion under analysis with the one which
has been created starting from the backedge.

This preprocessing phase is made in a fixed point fashion.